Add test implementations for TableStoreData spill

+++ server/core/src/main/scala/org/oliverlittle/clusterprocess/util/MemoryUsage.scala
@@ -2,7 +2,7 @@ package org.oliverlittle.clusterprocess.util
 
 // Adapted from: https://stackoverflow.com/questions/3571203/what-are-runtime-getruntime-totalmemory-and-freememory
 object MemoryUsage:
-    def getMemoryUsagePercentage(runtime : Runtime) : Double = MemoryUsage.getUsedMemory(runtime) / runtime.maxMemory
+    def getMemoryUsagePercentage(runtime : Runtime) : Double = MemoryUsage.getUsedMemory(runtime).toDouble / runtime.maxMemory.toDouble
 
     def getMemoryUsageAboveThreshold(runtime : Runtime, threshold : Double) : Long = {
         val usedPercentage = getMemoryUsagePercentage(runtime)
@@ -13,4 +13,4 @@ object MemoryUsage:
 
     def getTotalFreeMemory(runtime : Runtime) : Long = runtime.maxMemory - MemoryUsage.getUsedMemory(runtime).toLong
 
-    def getUsedMemory(runtime : Runtime) : Double = runtime.totalMemory - runtime.freeMemory
+    def getUsedMemory(runtime : Runtime) : Long = runtime.totalMemory - runtime.freeMemory
